Adição de mais recursos

+++ src/util/GoogleStorage.js
@@ -1,12 +1,9 @@
 const { Storage } = require("@google-cloud/storage");
 
-const storage = Boolean(process.env.DESENVOLVIMENTO)
-  ? new Storage({
+const storage =
+  new Storage({
     projectId: "universal-valve-275012",
     keyFilename: "D:/Downloads/universal-valve-275012-2220cb9ae931.json"
-  })
-  : new Storage({
-    projectId: "universal-valve-275012",
   })
   ;
 
@@ -18,14 +15,17 @@ module.exports = {
     const arquivo = bucket.file(nome);
     return await arquivo
       .save(buffer)
-      .catch(e =>{
+      .catch(e => {
         console.log(e)
       });
   },
 
   async apagar(nome) {
-    const arquivo = bucket.file(nome);
-    return await arquivo.delete().catch(e => console.log(e));
+    if (nome) {
+      const arquivo = bucket.file(nome);
+      return await arquivo.delete().catch(e => console.log(e));
+    }
+    return null;
   }
 
 }
